w3c-reports/12092014/efl/runned-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-play-expected.txt


FAIL invoking resource selection with play() assert_equals: networkState in separate script expected 0 but got 3(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-play.html:17:16
step@http://localhost:8800/resources/testharness.js:798:30
global code@http://localhost:8800/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-play.html:16:7)
